pyproject.toml
[project]
dependencies = [
  "hdbscan>=0.8.41",
  "json-repair>=0.53.1",
  "jupyter>=1.1.1",
  "langchain>=1.0.7",
  "langchain-community>=0.4.1",
  "langchain-openai>=1.0.3",
  "langgraph>=1.0.5",
  "matplotlib>=3.10.7",
  "pandas>=2.3.3",
  "pydantic>=2.12.4",
  "python-dotenv>=1.2.1",
  "scikit-learn>=1.8.0",
  "seaborn>=0.13.2",
  "sentence-transformers>=5.2.0",
]
description = "Add your description here"
name = "theory-autoencoder"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
